requirements.txt
fastapi
uvicorn
pydantic
anthropic
pandas
numpy
openpyxl
plotly
python-dotenv
ucimlrepo
kaleido
weasyprint
